adding the user voucher controller
